ref(RefactorStructure): refactor on custom fsd methodology
