+++ src/components/PrivateRoute.tsx
@@ -1,6 +1,6 @@
 import { Navigate, useLocation } from "react-router-dom";
-import { useAuthContext } from "../hooks/useAuthContext";
-import { publickRoutes } from "../routes/routes";
+import { useAuthContext } from "../shared/hooks/useAuthContext";
+import { publickRoutes } from "../shared/config/routes.config";
 
 export const PrivateRoute = ({children}: React.PropsWithChildren) => {
     const auth = useAuthContext();
