+++ src/context/AuthProvider/AuthContext.ts
@@ -1,4 +1,4 @@
 import { createContext } from "react";
-import type { IAuthContextType } from "../../types/context.types";
+import type { IAuthContextType } from "../../shared/types/context.types";
 
 export const AuthContext = createContext<IAuthContextType | null>(null);
