+++ src/shared/hooks/useAuthContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { AuthContext } from "../context/AuthProvider/AuthContext";
+import { AuthContext } from "../../context/AuthProvider/AuthContext";
 
 export const useAuthContext = () => {
     return useContext(AuthContext);
