+++ src/shared/hooks/useIntersectionObserver.ts
