+++ src/shared/types/logic.types.ts
