+++ src/shared/types/ui.types.ts
